Support Python versions before v2.6
The 'io' module was introduced in v2.6.  Create a helper function
that will read the first line from a file and return it.  This
fits all but one of the users of io.FileIO.  The exception is the
parser for /proc/mounts.  For it, open-code the use of open()
and readline() to replace io.FileIO.