Correct indentation when printing two-line descriptions
show_block() prints two-line descriptions spaced to align the text of
line #2 with the text after the block node on line #1.  When a the
block device had children, the spacing was off, as the utf-8 code for
the vertical bar is three bytes long.  Arguably, the code should have
all of its strings converted to u'' notation, but the simpler fix is
to always use the length of the plain spacer when computing the indent.