Simplify probe_block() by using the Struct() attribute ==> file lookup capability.
The Struct() class was previously extended to automatically look up unknown
attributes in the filesystem if 'attrpath' is defined.  Restructure probe_block()
to leverage this feature to simplify the code.  This section reduces the need to
call out to mdadm for active arrays.

While editing probe_block, rename the function-local variable 'dev' to 'blk' for
consistency with other functions.